implemented freight tracker module
