+++ assets/popup.css
@@ -82,7 +82,7 @@ a {
 .margin-12 {
   margin: 12px;
 }
-.mt-16 {
+.lcz-mt-16 {
   margin-top: 16px;
 }
 .lcz-mb-16 {
