+++ src/popup-content.js
@@ -7,6 +7,8 @@ import { isDomainValid } from "./content";
 import { detectPhoneModel } from "./phone-utils";
 import { getPhoneCarbonData } from "./phone-utils";
 import { getRecommendedModels } from "./phone-utils";
+import { getFedexDataChange, observeFedexShippingOptions,
+  recordAllInputChange, recordPackageTypeChange, recordFromToAddressChange } from "./autodetect/freight/freight-tracker";
 
 const LCA_SERVER_URL = "https://lca-server-api.fly.dev";
 
@@ -1204,7 +1206,7 @@ function trackFreight() {
   if (isDomainValid(allowedDomains)) {
     // observeFedexBtn();
     observeFedexShippingOptions(() => {
-      handleFedexDataToFreight();
+      handleFedexDataChange();
       recordAllInputChange();
       recordPackageTypeChange();
       recordFromToAddressChange();
@@ -1255,187 +1257,37 @@ function observeTextChange(element, callback) {
   observer.observe(element, { childList: true, subtree: true });
 }
 
-// Observes when the different shipping option appears
-function observeFedexShippingOptions(callback) {
-  const observer = new MutationObserver((mutationsList, observer) => {
-    for (const mutation of mutationsList) {
-      if (mutation.type === "childList") {
-        const shippingOption = document.querySelector(
-          ".fdx-c-definitionlist__description--small"
-        );
-        if (shippingOption) {
-          observer.disconnect(); // Stop observing once the shipping option is found
-          callback();
-          break;
-        }
-      }
-    }
-  });
-  observer.observe(document.body, { childList: true, subtree: true });
-}
-
-// Detects if an element's class has been changed.
-function onClassChange(element, callback) {
-  const observer = new MutationObserver((mutations) => {
-    mutations.forEach((mutation) => {
-      if (
-        mutation.type === "attributes" &&
-        mutation.attributeName === "class"
-      ) {
-        callback(mutation.target);
-      }
-    });
-  });
-  observer.observe(element, { attributes: true });
-  return observer.disconnect;
-}
-
-// Records the change of the from and to address
-function recordFromToAddressChange() {
-  const fromAddressElement = document.getElementById("fromGoogleAddress");
-  const toAddressElement = document.getElementById("toGoogleAddress");
-  // Listen to class changes on fromAddressElement and toAddressElement
-
-  const checkBothValid = () => {
-    const isFromValid = fromAddressElement.classList.contains("ng-valid");
-    const isToValid = toAddressElement.classList.contains("ng-valid");
-    if (isFromValid && isToValid) {
-      recordAllInputChange();
-      recordPackageTypeChange();
-      handleFedexChange();
-    }
-  };
-
-  onClassChange(fromAddressElement, checkBothValid);
-  onClassChange(toAddressElement, checkBothValid);
-}
-
-// Records the change of the package type
-function recordPackageTypeChange() {
-  const packageType = document.getElementById("package-details__package-type");
-  if (packageType) {
-    packageType.addEventListener("change", () => {
-      const packageWeightElement = document.getElementById(
-        "package-details__weight-0"
-      );
-      const packageCountElement = document.getElementById(
-        "package-details__quantity-0"
-      );
-
-      packageWeightElement.addEventListener("input", handleFedexChange);
-      packageCountElement.addEventListener("input", handleFedexChange);
-    });
-  }
-}
-
 // Observes the change of the shipping options
-function observeAndStoreShippingOptions() {
+export function observeAndStoreShippingOptions() {
   observeFedexShippingOptions(() => {
     console.log(
       "*****************OBSERVE FEDEX SHIPPING OPTIONS*****************"
     );
-    handleFedexDataToFreight();
+    handleFedexDataChange();
   });
 }
 
 // Handles the change of the shipping options
-function handleFedexChange() {
-  // console.log(`Changed value in ${event.target.tagName}:`, event.target.value);
-  const fedexButton = document.getElementById(
-    "e2ePackageDetailsSubmitButtonRates"
+async function handleFedexDataChange() {
+  let {fromAddress, toAddress, packageCount, packageWeight, currShippingOptions} = getFedexDataChange();
+  const totalWeight = packageWeight * packageCount;
+  const freightData = await getFreightData(
+    fromAddress,
+    toAddress,
+    totalWeight,
+    currShippingOptions
   );
-  if (fedexButton && !fedexButton.classList.contains("lca-viz-observing")) {
-    // Uses addEvent instead of addEventListener in order to ensure that we cannot add multiple event listeners
-    // addEvent will not add the same function twice.
-    fedexButton.addEventListener("click", () => {
-      observeAndStoreShippingOptions();
-    });
-    fedexButton.classList.add("lca-viz-observing");
-  }
-}
-
-function recordAllInputChange() {
-  // Select all input, select, and textarea elements
-  const inputs = document.querySelectorAll("input, select, textarea");
-
-  // Add event listeners to all selected elements
-  inputs.forEach((input) => {
-    if (
-      input.id !== "package-details__package-type" &&
-      input.id !== "fromGoogleAddress" &&
-      input.id !== "toGoogleAddress"
-    ) {
-      input.addEventListener("change", handleFedexChange);
-      input.addEventListener("input", handleFedexChange);
-    }
-  });
-}
-
-// Handles the change of the shipping options
-async function handleFedexDataToFreight() {
-  let currShippingOptions = [];
-  const availableOptions = document.querySelectorAll(
-    ".fdx-c-definitionlist__description--small"
-  );
-  availableOptions.forEach((option) => {
-    currShippingOptions.push(option.outerText.toLowerCase().replace(/®/g, ""));
-  });
-
-  const fromAddressElement = document.getElementById("fromGoogleAddress");
-  const fromAddress = fromAddressElement ? fromAddressElement.value : null;
-
-  const toAddressElement = document.getElementById("toGoogleAddress");
-  const toAddress = toAddressElement ? toAddressElement.value : null;
-
-  const packageCountElement = document.getElementById(
-    "package-details__quantity-0"
-  );
-  const packageCount = packageCountElement
-    ? parseInt(packageCountElement.value)
-    : null;
-
-  // const packageWeightElement = document.getElementById("package-details__weight-0");
-  const packageWeightElement = document.querySelector(
-    "#package-details__weight-0 .fdx-c-form__input"
-  );
-  let packageWeight = packageWeightElement
-    ? parseInt(packageWeightElement.value)
-    : null;
-
-  const unitElement = document.querySelector(
-    'select[data-e2e-id="selectMeasurement"]'
-  );
-  const unit = unitElement ? unitElement.value : null;
-
-  // If the unit is imperial, convert the package weight to kg.
-  if (unit.includes("IMPERIAL")) {
-    packageWeight = toKg(packageWeight);
-  }
-
-  if (fromAddress && toAddress && packageCount && packageWeight) {
-    const totalWeight = packageWeight * packageCount;
-    const freightData = await getFreightData(
-      fromAddress,
-      toAddress,
-      totalWeight,
-      currShippingOptions
-    );
-    if (shadowRoot.querySelector(".freight-container") !== null) {
-      console.log("updating freight content popup.....");
-      await updateFreightContent(freightData);
-    } else {
-      console.log("injecting freight content popup.....");
-      await injectPopupContent("freight", freightData);
-    }
-    currShippingOptions = [];
+  console.log("freightData: ", freightData);
+  if (shadowRoot.querySelector(".freight-container") !== null) {
+    console.log("updating freight content popup.....");
+    await updateFreightContent(freightData);
   } else {
-    console.log("fromAddress: ", fromAddress);
-    console.log("toAddress: ", toAddress);
-    console.log("packageCount: ", packageCount);
-    console.log("packageWeight: ", packageWeight);
-    console.error("Invalid input.. Information is not complete");
+    console.log("injecting freight content popup.....");
+    await injectPopupContent("freight", freightData);
   }
+  currShippingOptions = [];
 }
+
 
 // Loads the Google maps
 async function loadGoogleMaps(freightAirData, freightGroundData) {
@@ -1499,11 +1351,6 @@ function injectGoogleMaps(mode) {
   iframe.scrolling = "no";
 
   mapsContainer.appendChild(iframe);
-}
-
-// Converts the weight from lbs to kg
-function toKg(lbs) {
-  return lbs * 0.453;
 }
 
 // Handles searching for a phone model from the database
