+++ src/utils/math-utils.js
@@ -1,0 +1,5 @@
+
+// Utilities function relating to calculation
+export function LbtoKg(lbs) {
+  return lbs * 0.453;
+}
